chore: 확장함수 공통 사용을 위해 모듈 이동
