+++ android/feature/invitation/src/main/kotlin/com/andlife/invitation/model/guestbook/collection/InvitationCollectionUiModel.kt
@@ -1,7 +1,7 @@
 package com.andlife.invitation.model.guestbook.collection
 
 import com.andlife.domain.model.guestbook.GalleryMedia
-import com.andlife.invitation.util.toUiType
+import com.andlife.model.util.toUiType
 import com.andlife.model.guestbook.UiMediaType
 import kotlinx.datetime.LocalDateTime
