+++ android/feature/invitation/src/main/kotlin/com/andlife/invitation/screen/guestbook/InvitationCollectionScreen.kt
@@ -14,7 +14,7 @@ import com.andlife.invitation.component.InvitationMediaGridView
 import com.andlife.invitation.model.guestbook.collection.InvitationCollectionUiEvent
 import com.andlife.invitation.model.guestbook.collection.InvitationCollectionUiModel
 import com.andlife.invitation.model.guestbook.collection.InvitationCollectionUiState
-import com.andlife.invitation.util.toUiType
+import com.andlife.model.util.toUiType
 import com.andlife.invitation.viewmodel.InvitationCollectionViewModel
 import kotlinx.collections.immutable.persistentListOf
 import kotlinx.datetime.Clock
