+++ android/feature/model/src/main/kotlin/com/andlife/model/util/MediaTypeExtensions.kt
@@ -1,10 +1,10 @@
-package com.andlife.invitation.util
+package com.andlife.model.util
 
 import com.andlife.domain.model.guestbook.MediaType
 import com.andlife.model.guestbook.UiMediaType
 
 /**
- * Domain의 MediaType을 UI의 UiMediaType으로 변환하는 확장 함수
+ * Domain의 MediaType을 UI의 UiMediaType으로 변환하는 매퍼 확장 함수
  */
 fun MediaType.toUiType(): UiMediaType =
     when (this) {
